backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.29.0
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.9.2